Завершает фильтрацию данных
