+++ js/util.js
@@ -28,7 +28,7 @@ const getRandomArrayElement = (elements) => elements[getRandomInteger(0, element
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
-function debounce (callback, timeoutDelay = 5000) {
+function debounce (callback, timeoutDelay = 500) {
 
   let timeoutId;
